tests/mse_webm/big.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: Add big data</title>
  <script type="text/javascript" src="../mediasource.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

// This test checks loading a stereo segment, followed by a 5.1 segment plays without error.

runWithMSE(function(ms, el) {
  el.controls = true;
  once(ms, 'sourceopen').then(async function() {
    // Log events for debugging.
    var events = ["suspend", "play", "canplay", "canplaythrough", "loadstart", "loadedmetadata",
                  "loadeddata", "playing", "ended", "error", "stalled", "emptied", "abort",
                  "waiting", "pause", "durationchange", "seeking", "seeked"];
    function logEvent(e) {
      var v = e.target;
      info("got " + e.type + " event");
    }
    events.forEach(function(e) {
      el.addEventListener(e, logEvent);
    });

    info("MediaSource.isTypeSupported('video/webm; codecs=vp9,opus') = " + MediaSource.isTypeSupported('video/webm; codecs=vp9,opus'));
    ok(true, "Receive a sourceopen event");
    var sb = ms.addSourceBuffer("video/webm; codecs=vp9,opus");
    el.addEventListener("error", function(e) {
      ok(false, "should not fire '" + e.type + "' event");
      SimpleTest.finish();
    });
    is(el.readyState, el.HAVE_NOTHING, "readyState is HAVE_NOTHING");
    var promises = [];
    const buffer = await fetchWithXHR("../../mediatest/webm/Morocco_1080p60.webm");
    const initSegmentSize = 722;
    const cueSegmentOffset = buffer.byteLength; //98591072;
    var p1 = loadSegment(sb, buffer.slice(0, initSegmentSize));
    var p2 = once(el, "loadedmetadata");
    await Promise.all([p1, p2]);
    var total = 0;
    ms.duration = 1000; // to remove the thousands of durationchange events with safari
    let offset = initSegmentSize;
    const lengthToAdd = 1000000;
    for (;;) {
        let offset_end = offset + lengthToAdd;
        if (offset_end > cueSegmentOffset)
            offset_end = cueSegmentOffset;
        try {
            await loadSegment(sb, buffer.slice(offset, offset_end));
        } catch(err) {
            info("run out of room at " + sb.buffered.end(sb.buffered.length-1) + " after " + total + " bytes");
            break;
        }
        total += offset_end - offset;
        offset = offset_end;
        info(total + " bytes added");
        if (offset_end >= cueSegmentOffset) {
            offset = 0; // initSegmentSize;
            const endRange = sb.buffered.end(sb.buffered.length-1);
            sb.timestampOffset = endRange;
        }
    }
    ms.endOfStream();
    return once(el, 'ended');
  })
  .then(function() {
    ok(el.currentTime >= el.buffered.end(0), "played to the end");
    SimpleTest.finish();
  });
});

</script>
</pre>
</body>
</html>
